% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{bbb_search}
\alias{bbb_search}
\title{Search the BBB ScamTracker for incidents}
\usage{
bbb_search(keywords = NULL, scam_types = NULL, country = c("CAN-USA",
  "CAN", "USA"), state = NULL, date_start = Sys.Date(),
  date_end = Sys.Date())
}
\arguments{
\item{keywords}{keywords}

\item{scam_types}{scam_types}

\item{country}{country}

\item{date_start, date_end}{search date start/end (defaults to today)}
}
\description{
NOTE: Search API only returns max of 500 results.
}
