% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/details.r
\name{scam_details}
\alias{scam_details}
\title{Retrive details about a BBB spam report}
\usage{
scam_details(pkid = NULL)
}
\arguments{
\item{pkid}{pkid}
}
\description{
Retrive details about a BBB spam report
}
