% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.r
\docType{data}
\name{scam_types}
\alias{scam_types}
\title{Scam types}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 31 rows and 2 columns.}
\usage{
scam_types
}
\description{
Scam types
}
\keyword{datasets}
