% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scamtracker-package.R
\docType{package}
\name{scamtracker}
\alias{scamtracker}
\alias{scamtracker-package}
\title{R interface to the BBB ScamTracker}
\description{
Website: \url{https://www.bbb.org/scamtracker/us}
}
\author{
Bob Rudis (@hrbrmstr)
}

