% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.r
\name{bbb_search}
\alias{bbb_search}
\title{Search the BBB ScamTracker for incidents}
\usage{
bbb_search(scam_types = NULL, scam_name = NULL, date_start = NULL,
  date_end = NULL)
}
\arguments{
\item{scam_types}{scam_types}

\item{scam_name}{scam_name}

\item{date_start}{date_start}

\item{date_end}{date_end}
}
\description{
Search the BBB ScamTracker for incidents
}

